git-svn-id: file:///Users/braun/svn/vermont/trunk/topas@1543 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ detectionmodules/statmodules/wkp-module/pcs-test.cpp
@@ -216,10 +216,11 @@ void pcs_categories ( std::list<int64_t> sample1, std::list<int64_t> sample2,
   unsigned lg2 = sample2.size();
   unsigned n = lg1 + lg2;
 
-  unsigned min = sample1.front()<sample2.front() ? sample1.front():sample2.front();
+  int64_t min = sample1.front()<sample2.front() ? sample1.front():sample2.front();
   unsigned sample1_oldsize = lg1;
   unsigned sample2_oldsize = lg2;
   double E1, E2;
+
   // M.L.E. of the probability that an item in sample1 (E1) and sample2 (E2)
   // will be in the current category, i.e. expected number of observations
   // (of items, of elements) in the current category for both samples;
